Fix but in redis parser and clean up a bit.
The parser would never had parsed a null array because the check
looked for the entire line ("*-1\r\n"), not just the length ("-1").
Added a test for the null Array case.

Cleaned up bits at the beginning to not make intermediate collections.

Reformatted the `parse` function to be expression oriented as opposed
to statement oriented.