Parsing does dramatically less copying
The actual parsing process doesn't need a unique copy of the buffer so it can actually be shared safely. Places that required a moved Vec<u8> can now take a &[u8] and places that required a moved String now take a &str, all of which are just views of the ByteBuf contents.