Reduce copies in thrift buffer system
The `Buffer` struct contains an inner Vec<u8> which acts as the real
buffer. When extracting it, there is almost never a need to keep the
origional buffer, so simply moving its inner `Vec` out is the best
course of action. To make up for situations where one would actually
want to keep the `Buffer` around, it now derives the `Clone` trait.