Merge workload and request crates
closes #72

workload had very little reason for being its own crate so it was
merged with the request module.

Remove `ThriftType` tree.
We now just use `Parameter` and its field `Value`. `ThriftType` was
as subset of the information already stored in the vector of
`Parameter`s so it make sense to just use it.

Add support for type `double` in thrift

Removed `ThriftRequest` type