Require a config file
This was already the case, enforced in the main function when it checks
for a non-zero number of workloads. Workloads are only generated based on
a config file so implicitly a config file is required. However, config
loading was happy to let you not have one.

Now config loading results in an error if no config file is present and
the check on the number of workloads is also in the config file.