Reimplement the notion of a protocol as a trait
Problem: The concept of a `protocol` is currently satisfied by an `enum`. This
causes handling of different protocols to be strongly coupled which is fine
for protocols that are very similar such as redis and memcached but begins to
cause fiction when trying to accommodate very different protocols such as
thrift.

Solution: Change the concept of a protocol to be enshrined by a `trait`. This
leads to a number of advantages:
* The concept of a protocol is 'open' by nature
* A generic interface can be presented that hides the protocol details from
  the network and load management details.
* Compilation units can be split into different crates to make building faster.

Details
Protocol message generation is enshrined in the trait `ProtocolGen`. Its only
purpose is to generate buffers and can identify itself with a text string. A
single instance of a `ProtocolGen` is needed so there is no need for a factory
type.

Protocol message parsing is handled by the `ProtocolParse` trait. Its only
method enshrines parsing a buffer into a `ProtocolResponse`. The current model
creates many threads to parse messages so this requires a factory method to
make new `ProtocolParse` instances when necessary.

The type `Parameter` has also been generalized. Its value is abstract and
enshrined by the trait `Ptype` which dictates a way to parse a new value and
regenerate a random version of itself.

Each protocol gets a new crate which only needs to provide a single function
which is capable of loading its configuration, creating its workload instances
and a `ProtocolParseFactory` for spawning new parsing instances.