datastructures: histogram improvements (#202)
Improves the performance of the percentile operations for the
histogram datastructures by using slightly more memory to maintain
a secondary set of counts that cover wider ranges. This is used to
accelerate the totals() and percentile() functions.

Adds a new CircularHistogram which is like MovingHistogram
but uses a fixed-size circular buffer to store the sample history. It is
faster than MovingHistogram and suitable for cases where it is
possible to know in-advance an upper-bound on the number of
samples across the integration window. One example would be for
storing samples at fixed intervals.