refactored atomics and datastructures
Adds a new atomics library which provides nicer generic atomics.

Refactors the datastructures, using a single Histogram implementation
which can accomodate use-cases with or without a moving window and
with or without a fixed sample capacity. `Counter`s are also
significantly refactored

Changes to other libraries and rpc-perf to support the changes elsewhere