metrics: change behavior of clear() (#210)
Previously, clear() would just zero out the data stored in the
recorder without removing channels that have been registered.

With this change, clear() behaves like it does on a HashMap and
actually removes registered channels. The new function, zero(),
can be used to zero the data stored, without deregistering all
of the channels.