refactoring to reduce memory utilization (#212)
Major refactoring of atomics, datastructures, and metrics to help
reduce memory utilization. Previously, we were using `Box<>`'s
heavily to help make the datastructures more generic. This came at
a high cost for the histogram implementation, as we had a `Box<>`
for every counter in the histogram.

With this change, we refactor to remove the boxes but still allow
for the datastructures and metrics to be generic in terms of width.
By doing this, we significantly reduce the memory footprint.