metrics: refactor to reduce complexity
Refactors the metrics library to reduce complexity. The difference
between the `Metrics` and `Recorder` structs was unnecessary. Instead,
we are able to make it a single struct.

Adds a `Delta` measurement type that allows for incrementing a counter
by some value. This is helpful for aggregating metrics from things like
individual requests.