waterfall: fix rendering issues for some distributions (#279)
Previously, certain distributions could result in an all black
output. This was caused by rounding issues with integer arithmetic
in calculating the weighted count for the buckets.

Adjust the weighting algorithm to use a fixed multiplier and
increase precision used for storing the weighted counts.
Additionally, alter the coloring algorithm to improve rendering of
wide distributions. Adds additional distributions to the simulator
which were used to help identify issues.

The result is a more robust waterfall rendering which handles a
wider variety of distributions.