datastructures: bugfix for moving and windowed histograms
An issue occurs when bounded or windowed histograms have a value
larger than the max incremented or decremented. A sample was not
properly added to the queue for trimming in the future, this would
result in the datastructure accumulating a count beyond the limit
of the histogram. When enough of these samples have been seen, the
histogram would report the max value for almost all percentiles.

Added some tests that check for the expected behavior and fixed the
logic in the increment and decrement functions.