simplify client and session modules and update mio
Simplifies the client and session modules by unifying the TLS and
non-TLS implementations. This made it possible to address some
sublte bugs in the state machine and readiness model when running
against a TLS enabled pingserver. Additionally, this made the mio
upgrade much easier, as it reduces the code complexity. This
benefit will carry-forward to future upgrades and enhancements.