fix: prometheus metric naming and typing (#317)
Fixes prometheus metric naming and adds type annotations.

Prometheus metric names cannot contain a '.' which was previously used in the
percentile metrics. This changes the percentile to be exposed with a different
label value for each percentile.

OpenTelemetry requires metrics to have prometheus type annotations. Adds
basic type annotations to the metrics so that the metrics can be collected by
OpenTelemetry.