Add cardinality to U32 key/value configs
The cardinality of the keyspace being constant (in the case of the U32
field type) or tied directly to the 52^N of the keyspace length (in the
case of the Alphanumeric field type) makes it difficult to control the
exact size of the keyspace:

len1: 52
len2: 2,704
len3: 140,608
len4: 7,311,616
len5: 380,204,032
len6: 19,770,609,664

The wide gaps between these keyspace cardinalities make it difficult to
design load test configurations that achieve a specific hit-rate.

In order to solve this problem, this commit implements the "cardinality"
config option for key/innerkey/value spaces for U32 field types which
constrains the range of the random number generator used to generate
these fields.